requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
ca-certificates=2021.10.8=h5b45459_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
certifi=2021.10.8=py39hcbf5309_0
cloudpickle=1.6.0=py_0
colorama=0.4.4=pyh9f0ad1d_0
cycler=0.10.0=py_2
decorator=5.0.7=pyhd8ed1ab_0
freetype=2.10.4=h546665d_1
h5py=3.2.1=nompi_py39hf27771d_100
hdf5=1.10.6=nompi_h5268f04_1114
icu=68.1=h0e60522_0
intel-openmp=2021.2.0=h57928b3_616
ipykernel=5.5.3=py39h832f523_0
ipython=7.22.0=py39h832f523_0
ipython_genutils=0.2.0=py_1
jedi=0.18.0=py39hcbf5309_2
jpeg=9d=h8ffe710_0
jupyter_client=6.1.12=pyhd8ed1ab_0
jupyter_core=4.7.1=py39hcbf5309_0
kiwisolver=1.3.1=py39h2e07f2f_1
krb5=1.17.2=hbae68bd_0
lcms2=2.12=h2a16943_0
libblas=3.9.0=8_mkl
libcblas=3.9.0=8_mkl
libclang=11.1.0=default_h5c34c98_0
libcurl=7.76.1=hf1763fc_1
liblapack=3.9.0=8_mkl
libpng=1.6.37=h1d00b33_2
libsodium=1.0.18=h8d14728_1
libssh2=1.9.0=h680486a_6
libtiff=4.2.0=hc10be44_1
lz4-c=1.9.3=h8ffe710_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matplotlib=3.4.1=py39hcbf5309_0
mkl=2020.4=hb70f87d_311
msys2-conda-epoch=20160418=1
nibabel=3.2.1=pyhd8ed1ab_0
numpy=1.20.2=py39h6635163_0
olefile=0.46=pyh9f0ad1d_1
opencv-python=4.5.3.56=pypi_0
openjpeg=2.4.0=h48faf41_0
openssl=1.1.1l=h8ffe710_0
packaging=20.9=pyh44b312d_0
parso=0.8.2=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=8.1.2=py39h1a9d4f7_1
pip=21.0.1=pyhd8ed1ab_0
prompt-toolkit=3.0.18=pyha770c72_0
pydicom=2.1.2=pyhd3deb0d_0
pygments=2.8.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py39hcbf5309_7
pyqt-impl=5.12.3=py39h415ef7b_7
pyqt5-sip=4.19.18=py39h415ef7b_7
pyqtchart=5.12=py39h415ef7b_7
pyqtgraph=0.12.3=pyhd8ed1ab_0
pyqtwebengine=5.12.1=py39h415ef7b_7
pyreadline=2.1=py39hcbf5309_1003
python=3.9.2=h7840368_0_cpython
python-dateutil=2.8.1=py_0
python_abi=3.9=1_cp39
pywin32=300=py39hb82d6ee_0
pyzmq=22.0.3=py39he46f08e_1
qt=5.12.9=h5909a2a_4
scipy=1.6.2=py39hc0c34ad_0
setuptools=49.6.0=py39hcbf5309_3
six=1.15.0=pyh9f0ad1d_0
spyder-kernels=2.0.1=py39hcbf5309_0
sqlite=3.35.5=h8ffe710_0
tk=8.6.10=h8ffe710_1
tornado=6.1=py39hb82d6ee_1
traitlets=5.0.5=py_0
tzdata=2021a=he74cb21_0
vc=14.2=hb210afc_4
vs2015_runtime=14.28.29325=h5e1d092_4
wcwidth=0.2.5=pyh9f0ad1d_2
wheel=0.36.2=pyhd3deb0d_0
wincertstore=0.2=py39hcbf5309_1006
xz=5.2.5=h62dcd97_1
zeromq=4.3.4=h0e60522_0
zlib=1.2.11=vc14h1cdd9ab_1
zstd=1.4.9=h6255e5f_0
